ecommerce/apps/payments/urls.py
"""
URL patterns for the payments app.
"""
from django.urls import path, include
from rest_framework.routers import DefaultRouter

from .views import PaymentViewSet

app_name = 'payments'

router = DefaultRouter()
router.register('', PaymentViewSet, basename='payment')

urlpatterns = [
    path('', include(router.urls)),
]
