ecommerce/requirements.txt
# Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1

# API Documentation
drf-spectacular==0.27.0

# Async Tasks
celery==5.3.4
redis==5.0.1

# AI
anthropic>=0.39.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0

# Server
gunicorn==21.2.0

# Utils
python-dotenv==1.0.0
django-filter==23.5
